+++ docs/HISTORY.rst
@@ -7,6 +7,8 @@ Change Log
 **Features**
 
 **Fixes**
+- Fixed issue whereby `turn_on`, `turn_off` and `toggle` for some HASS entities didn't work
+- Fixed issue whereby `render_template` didn't work
 
 **Breaking Changes**